requirements.txt
collection
pathlib
typing
numpy
pillow
tqdm
torch
torchvision
nltk
pytest-warnings
matplotlib
pandas
ipython
gradio